rename : coverLetter/preview 구조 조정
