+++ client/src/features/coverletter/preview/ui/coverletter-list/CoverLetterItem.tsx
