+++ client/src/features/coverletter/preview/ui/coverletter-preview/CoverLetterPreView.tsx
