+++ client/src/pages/coverletter/new/index.tsx
@@ -6,13 +6,16 @@ import { CoverLetterCreator } from '@/src/features/coverletter/new';
 
 import { styleMixin } from '@/src/shared/styles';
 import { Spinner } from '@/src/shared/ui';
+import { APIErrorBoundary } from '@/src/features/boundary';
 
 function CoverLetterEditorPage() {
   return (
     <CoverLetterStyle>
-      <Suspense fallback={<Spinner />}>
-        <CoverLetterCreator />
-      </Suspense>
+      <APIErrorBoundary>
+        <Suspense fallback={<Spinner />}>
+          <CoverLetterCreator />
+        </Suspense>
+      </APIErrorBoundary>
     </CoverLetterStyle>
   );
 }
